refector : modify project arch
