--- admin.html
@@ -1,1 +1,0 @@
-<!DOCTYPE html><html lang="en"><head><meta charset="utf-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width,initial-scale=1"><link rel="icon" href="favicon.png"><title>Socket.IO Admin UI</title><link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900"><link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@mdi/font@latest/css/materialdesignicons.min.css"><link href="css/app.6e3b9661.css" rel="preload" as="style"><link href="css/chunk-vendors.9f55d012.css" rel="preload" as="style"><link href="js/app.0d7d7845.js" rel="preload" as="script"><link href="js/chunk-vendors.934c03ff.js" rel="preload" as="script"><link href="css/chunk-vendors.9f55d012.css" rel="stylesheet"><link href="css/app.6e3b9661.css" rel="stylesheet"></head><body><noscript><strong>We're sorry but Socket.IO Admin UI doesn't work properly without JavaScript enabled. Please enable it to continue.</strong></noscript><div id="app"></div><script src="js/chunk-vendors.934c03ff.js"></script><script src="js/app.0d7d7845.js"></script></body></html>
