+++ package.json
@@ -5,14 +5,15 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
+    "start": "node src/index.js"
   },
-  "type":"module",
+  "type": "module",
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@socket.io/admin-ui": "^0.5.1",
     "express": "^4.17.1",
+    "redis": "^4.6.14",
     "socket.io": "^3.0.4"
   }
 }
