+++ src/admin/favicon.png
